Added project files to github
The index page is the front page of the website, it contains the website name and the logo. Its content is in index.html and index.css.
The login page consists the username, password and signup option, the content is in the login.jsp, the signup content is in the signup.jsp and signup1.jsp.
When you login successfully, the product page appears on the screen. The product page contains on the products such as Books, pens, files, markers, pencils, erasers, scales, their code is in books.jsp, pens.jsp, files.jsp, markers.jsp, pencils.jsp, erasers.jsp, scales.jsp, respectively.
When user adds products to cart and clicks the cart button on the top-left of the page, the products are shown, the code for this is the cart.jsp.
The logout page content is in the logout.jsp. As soon as the logout option is clicked, the thankyou page appears and in few seconds the front page is shown again. The content of thankyou image code is in the logout.jsp.
The Admin after login, can check the no. of orders and details of all the customers. The customer details is in the alluserinfo.jsp. The admin can see the orders separately for all the products with Admin-books.jsp, Admin-pens.jsp, Admin-pencils.jsp, Admin-files.jsp, Admin-erasers.jsp, Admin-scales.jsp, Admin-markers.jsp respectively.